blog-series/02-strands-basics/requirements.txt
strands-agents
strands-agents[ollama]
strands-agents-tools
pydantic
